+++ member-service/src/main/java/com/hedvig/memberservice/identity/repository/IdentificationMethod.kt
@@ -1,0 +1,6 @@
+package com.hedvig.memberservice.identity.repository
+
+enum class IdentificationMethod {
+    NORWEGIAN_BANK_ID,
+    DANISH_BANK_ID
+}
