+++ member-service/src/main/java/com/hedvig/memberservice/identity/repository/IdentityRepository.kt
@@ -1,0 +1,5 @@
+package com.hedvig.memberservice.identity.repository
+
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface IdentityRepository : JpaRepository<IdentityEntity, Long>
