+++ member-service/src/main/java/com/hedvig/memberservice/identity/repository/Nationality.kt
@@ -1,0 +1,7 @@
+package com.hedvig.memberservice.identity.repository
+
+enum class Nationality {
+    SWEDEN,
+    NORWAY,
+    DENMARK
+}
